+++ nedsettings_form.php
@@ -66,9 +66,9 @@ class course_ned_edit_form extends moodleform {
         unset($activitytrackingbackgroundoptions);
 
         $locationoftrackingiconsoptions = array(
-            'moodleicons' => get_string('moodleicons', 'format_ned'),
-            'nediconsleft' => get_string('nediconsleft', 'format_ned'),
-            'nediconsright' => get_string('nediconsright', 'format_ned')
+            \format_ned\toolbox::$moodleicons => get_string('moodleicons', 'format_ned'),
+            \format_ned\toolbox::$nediconsleft => get_string('nediconsleft', 'format_ned'),
+            \format_ned\toolbox::$nediconsright => get_string('nediconsright', 'format_ned')
         );
         $mform->addElement('select', 'locationoftrackingicons',
             get_string('locationoftrackingicons', 'format_ned'), $locationoftrackingiconsoptions