+++ Changes.md
@@ -6,3 +6,4 @@ Version 3.3.0.1
      'course_section_cm_completion', thus if any theme extends the core course renderer then any changes will
      not be called for any method called directly or indirectly by the format renderer $this->courseformat attribute
      as effectivly they are two different objects.
+  3. Added 'nediconsleft' option.